fixed wrong DWORD, WORD conversions in logBackend/eventlog on Windows started to make /W2 the default, /W3 is too much for boost/asio/openssl

+++ src/logBackend.hpp
@@ -106,7 +106,7 @@ namespace _Wolframe {
 
 	private:
 		LogLevel::Level logLevel_;
-		DWORD categoryId_;
+		WORD categoryId_;
 		HANDLE eventSource_;
 		std::string log_;
 		std::string source_;
